requirements.txt
# Python 3.11 required
numpy==2.2.6
opencv-contrib-python==4.12.0.88
onnxruntime==1.23.0
PyYAML==6.0.2
scikit-learn==1.5.2
torch==2.8.0
torchvision==0.23.0
